PHRAS-1065 #time 1h backport PHRAS-947
