+++ lib/Alchemy/Phrasea/ControllerProvider/Api/V2.php
@@ -11,6 +11,7 @@ namespace Alchemy\Phrasea\ControllerProvider\Api;
 
 use Alchemy\Phrasea\Application as PhraseaApplication;
 use Alchemy\Phrasea\Controller\Api\BasketController;
+use Alchemy\Phrasea\Controller\Api\LazaretController;
 use Alchemy\Phrasea\Controller\Api\SearchController;
 use Alchemy\Phrasea\ControllerProvider\ControllerProviderTrait;
 use Alchemy\Phrasea\Core\Event\Listener\OAuthListener;
@@ -18,7 +19,6 @@ use Silex\Application;
 use Silex\Controller;
 use Silex\ControllerProviderInterface;
 use Silex\ServiceProviderInterface;
-use Symfony\Component\HttpFoundation\Request;
 
 class V2 implements ControllerProviderInterface, ServiceProviderInterface
 {
@@ -34,6 +34,12 @@ class V2 implements ControllerProviderInterface, ServiceProviderInterface
                     ->setDataboxLoggerLocator($app['phraseanet.logger'])
                     ->setDispatcher($app['dispatcher'])
                     ->setJsonBodyHelper($app['json.body_helper']);
+            }
+        );
+
+        $app['controller.api.v2.lazaret'] = $app->share(
+            function (PhraseaApplication $app) {
+                return (new LazaretController($app));
             }
         );
 
@@ -72,6 +78,17 @@ class V2 implements ControllerProviderInterface, ServiceProviderInterface
 
         $controllers->match('/search/', 'controller.api.v2.search:searchAction');
 
+        $controllers->delete('/quarantine/', 'controller.api.v2.lazaret:quarantineItemEmptyAction')
+            ->bind('api_v2_quarantine_empty');
+
+        $controller = $controllers->delete('/quarantine/item/{lazaret_id}/', 'controller.api.v2.lazaret:quarantineItemDeleteAction')
+            ->bind('api_v2_quarantine_item_delete');
+        $this->addQuarantineMiddleware($controller);
+
+        $controller = $controllers->post('/quarantine/item/{lazaret_id}/add/', 'controller.api.v2.lazaret:quarantineItemAddAction')
+            ->bind('api_v2_quarantine_item_add');
+        $this->addQuarantineMiddleware($controller);
+
         return $controllers;
     }
 
@@ -81,6 +98,14 @@ class V2 implements ControllerProviderInterface, ServiceProviderInterface
             ->before($app['middleware.basket.converter'])
             ->before($app['middleware.basket.user-access'])
             ->assert('basket', '\d+');
+
+        return $controller;
+    }
+
+    private function addQuarantineMiddleware(Controller $controller)
+    {
+        $controller
+            ->assert('lazaret_id', '\d+');
 
         return $controller;
     }
